Improved motor movement, added logging, improved communication protocol, begun testing calibration code

+++ firmware/cycloarm/serial.h
@@ -11,7 +11,7 @@ void serial_read();
 void serial_send_structure(byte *struct_pointer, uint8_t struct_length);
 void serial_read_structure(byte *struct_pointer, uint8_t struct_length);
 
-extern uint8_t serial_incoming_byte;
-extern uint8_t serial_peek_incoming_byte;
+extern uint16_t serial_incoming_byte;
+extern uint16_t serial_peek_incoming_byte;
 
 #endif
